Tmp: success run after refactor. About to change FiberTask interface.

+++ LIWEngine/Core/LIWFrame.h
@@ -4,7 +4,7 @@
 
 struct LIWFrameData {
 	const std::chrono::system_clock::time_point m_timePoint;
-	const double m_timeDelta{ 0 };
-	const double m_timeFromStart{ 0 };
+	const float m_timeDelta{ 0 };
+	const float m_timeFromStart{ 0 };
 	bool m_singnalEnd{ false };
 };
